requirements.txt
asgiref==3.11.0
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
cramjam==2.11.0
dataclasses==0.6
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-mongodb-backend==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
idna==3.11
passlib==1.7.4
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
python-snappy==0.7.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
winkerberos==0.12.2
